Finished the base game
The game now properly alternately places red and blue chips, detects a winning gamestate. It notifies the players when somebody has won and clears the board, and displays the overall score at all times. The overall score is reset when the menu is returned to. Now all that is left is to code an AI for the singleplayer mode.