Created the AI
I created an AI to play against the player in singleplayer, the AI determines its move by first determining all possible positions it could place a piece, then calculating the number of adjacent in-a-rows for each of those slots to determine which slot is the most highly contested, and then chooses that slot to place its piece. The bot consistently recognizes 3 in a rows to block them or secure a win.