Fixed main.c and the bot
I realized a few of the loops in my checkForWin function weren't initializing the counting variable 'a' back to 0, so it wasn't actually detecting wins properly. Also made it so that if placing a chip in a certain slot would allow the opponent to win, the bot will avoid placing a chip in that slot.